Create dialog functions and structs, and link form.h
Functions newdlg and deldlg are basically the same as newwin and delwin, except
that they take a custom data type, DIALOG, which holds necessary information.
field will be a reference to the form.h field that holds the input from the
dialog.

Add form.h include and link in makefile.